Moved listener interfaces to a proper folder
