+++ networklibrary/src/main/java/com/eis/communication/network/ListenersInterfaces/InviteListener.java
@@ -1,6 +1,8 @@
-package com.eis.communication.network;
+package com.eis.communication.network.ListenersInterfaces;
 
 import com.eis.communication.Peer;
+import com.eis.communication.network.FailReason;
+import com.eis.communication.network.NetworkManager;
 
 /**
  * Listener for sent invitations to join the network, often passed as parameter in {@link NetworkManager#invite(Peer, InviteListener)}.
