+++ networklibrary/src/main/java/com/eis/communication/network/ListenersInterfaces/JoinInvitationListener.java
@@ -1,4 +1,4 @@
-package com.eis.communication.network;
+package com.eis.communication.network.ListenersInterfaces;
 
 /**
  * Listener for invitations to join a network.
